+++ milu/src/parser/template.rs
@@ -1,18 +1,27 @@
 use nom::{
     branch::alt,
-    bytes::streaming::is_not,
+    bytes::streaming::{is_not, tag}, // Added tag for parse_template_element
     character::streaming::char,
-    combinator::{map, value, verify},
+    combinator::{cut, map, value, verify}, // Added cut
     error::{ContextError, FromExternalError, ParseError},
     multi::fold_many0,
-    sequence::{delimited, preceded},
-    IResult, Parser,
+    sequence::{delimited, preceded, terminated}, // Added terminated
+    IResult,
+    Parser,
 };
 
-use crate::{parser::string::parse_escaped_whitespace, script::Value};
+use super::rules::op_0; // Corrected path
+use super::string::parse_escaped_whitespace;
+use super::string::parse_unicode; // Corrected path
+use super::Span; // Corrected path
+use crate::script::Value; // Corrected path
 
-use super::string::parse_unicode;
-use super::Span;
+// This was likely defined in string.rs or a shared place. For now, defining it here.
+// If string.rs also needs it, it should be moved to a common module like parser/common.rs or parser/utils.rs
+// For now, to get template.rs to compile, I'm including a version of it.
+// The original `parse_unicode` was in `string.rs` as `pub(super) fn parse_unicode...`
+// which means it was callable from `parser/mod.rs` (the old `parser.rs`).
+// If `string.rs` is now a sibling module to `template.rs`, then `string::parse_unicode` would be the path.
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 enum StringFragment<'a> {
@@ -22,6 +31,9 @@ enum StringFragment<'a> {
     TemplateElement(Value),
 }
 
+// Copied from original parser.rs context, may need path adjustments for parse_unicode
+// Assuming parse_unicode is made public in string.rs or moved to a shared location.
+// For now, this might cause an error if string::parse_unicode isn't visible.
 fn parse_escaped_char<'a, E>(input: Span<'a>) -> IResult<Span<'a>, char, E>
 where
     E: ParseError<Span<'a>> + FromExternalError<Span<'a>, std::num::ParseIntError>,
@@ -29,7 +41,7 @@ where
     preceded(
         char('\\'),
         alt((
-            parse_unicode,
+            parse_unicode, // This needs to be resolvable
             value('\n', char('n')),
             value('\r', char('r')),
             value('\t', char('t')),
@@ -39,7 +51,7 @@ where
             value('/', char('/')),
             value('"', char('"')),
             value('$', char('$')),
-            value('`', char('`')), // Added for escaped backtick \`
+            value('`', char('`')),
         )),
     )
     .parse(input)
@@ -52,6 +64,7 @@ fn parse_template_literal<'a, E: ParseError<Span<'a>>>(
     verify(not_quote_slash, |s: &Span| !s.is_empty()).parse(input)
 }
 
+// Note: op_0 comes from rules.rs, so that needs to be correctly imported.
 fn parse_template_element<'a, E>(input: Span<'a>) -> IResult<Span<'a>, Value, E>
 where
     E: ParseError<Span<'a>>
@@ -59,10 +72,9 @@ where
         + ContextError<Span<'a>>
         + std::fmt::Debug,
 {
-    // Use tag to match "${" literally. If this passes, then cut.
     preceded(
-        nom::bytes::streaming::tag("${"),
-        nom::combinator::cut(nom::sequence::terminated(super::op_0, char('}'))),
+        tag("${"),                        // Using nom::bytes::streaming::tag
+        cut(terminated(op_0, char('}'))), // op_0 needs to be in scope
     )
     .parse(input)
 }
@@ -75,11 +87,11 @@ where
         + std::fmt::Debug,
 {
     alt((
-        map(parse_escaped_char, StringFragment::EscapedChar), // Priority 1: Escaped characters (e.g., \$, \`, \\)
-        map(parse_template_element, StringFragment::TemplateElement), // Priority 2: Template expressions (e.g., ${expr})
-        map(char('$'), |c: char| StringFragment::EscapedChar(c)), // Priority 3: Literal '$', treated like an escaped char for simplicity
-        map(parse_template_literal, StringFragment::Literal), // Priority 4: Regular text segments
-        value(StringFragment::EscapedWS, parse_escaped_whitespace), // Priority 5: Escaped whitespace
+        map(parse_escaped_char, StringFragment::EscapedChar),
+        map(parse_template_element, StringFragment::TemplateElement),
+        map(char('$'), |c: char| StringFragment::EscapedChar(c)),
+        map(parse_template_literal, StringFragment::Literal),
+        value(StringFragment::EscapedWS, parse_escaped_whitespace), // This needs to be resolvable
     ))
     .parse(input)
 }
@@ -96,10 +108,10 @@ where
         Vec::<Value>::new,
         |mut acc, fragment| {
             match fragment {
-                StringFragment::Literal(s) => acc.push(s.to_string().into()),
+                StringFragment::Literal(s) => acc.push(s.fragment().to_string().into()), // .fragment() added
                 StringFragment::EscapedChar(c) => acc.push(c.to_string().into()),
                 StringFragment::TemplateElement(v) => acc.push(v),
-                _ => {}
+                StringFragment::EscapedWS => {} // Explicitly handle EscapedWS if it shouldn't add to vec
             }
             acc
         },
@@ -111,33 +123,31 @@ where
 mod tests {
     use std::sync::Arc;
 
-    use super::*;
+    use super::*; // For parse_template itself and Span from current module (template.rs)
     use crate::parser::parse; // For assert_parse_template_error helper
-    use crate::parser::tests::{array, id, int, plus, str, strcat}; // Adjust if macros are not public
-    use crate::script::stdlib::*;
-    use crate::script::Call;
-    use crate::script::Value;
+    use crate::parser::Span as ParserSpan; // Keep Span alias if needed, or use crate::parser::Span directly
+    use crate::script::stdlib::{Greater, If}; // For complex expression test
+    use crate::script::Value; // For complex expression test
+    use crate::{array, id, int, plus, str, strcat}; // Changed to crate:: for macros
+                                                    // For creating Spans in tests if needed, or use ParserSpan
 
     // Helper for asserting successful template parsing
     fn assert_template_ast(input: &str, expected_vec: Vec<Value>) {
-        let result = parse_template::<nom_language::error::VerboseError<Span>>(Span::new(input));
+        // Ensure Span type is consistent. parse_template expects nom_locate::LocatedSpan<&str>
+        // The Span type alias in parser/mod.rs should be this.
+        // If tests create Spans, they should use the same type.
+        let span_input = ParserSpan::new(input); // Use the aliased Span from parser
+        let result = parse_template::<nom_language::error::VerboseError<ParserSpan>>(span_input);
         assert!(result.is_ok(), "Parsing failed for input: {}", input);
         let (_, parsed_vec) = result.unwrap();
 
-        // Construct the expected Value::OpCall(Arc<Call>) for strcat
         let expected_value = if parsed_vec.is_empty() {
-            strcat!(array!()) // Handles `` ` `` resulting in strcat!([])
+            strcat!(array!())
         } else {
             strcat!(Value::Array(Arc::new(expected_vec)))
         };
 
-        // The parse_template function itself returns Vec<Value>,
-        // but the overall parser (in parser.rs) wraps this into a strcat call.
-        // To test template.rs in isolation, we check the Vec<Value>.
-        // However, for integration, it's good to also see how it fits.
-        // The main `parse` function (from parser.rs) does the strcat wrapping.
-        // Let's use that for a more integrated test of template parsing.
-        let overall_parse_result = crate::parser::parse(input); // Assuming parse() is accessible
+        let overall_parse_result = crate::parser::parse(input);
         assert!(
             overall_parse_result.is_ok(),
             "Overall parsing failed for input: {}. Error: {:?}",
@@ -153,18 +163,12 @@ mod tests {
         );
     }
 
-    // Helper for error tests
     fn assert_parse_template_error(input: &str) {
-        // We test errors through the main parse function because template parsing errors
-        // are often caught by the surrounding grammar or expression parsing within ${}.
         assert!(parse(input).is_err(), "Expected error for input: {}", input);
     }
 
-    // Basic Template Tests
     #[test]
     fn test_empty_template() {
-        // The parser wraps the result of parse_template (which is Vec<Value>)
-        // into a StringConcat call. So, `` ` `` becomes strcat!(array!()).
         assert_template_ast("``", vec![]);
     }
 
@@ -180,9 +184,6 @@ mod tests {
 
     #[test]
     fn test_template_with_multiple_expressions() {
-        // Note: The current parse_template logic might produce a flat Vec<Value>
-        // like vec![expr1, str!(" "), expr2]. The strcat macro in parser.rs tests might handle this.
-        // Let's assume spaces between expressions become separate string literals.
         assert_template_ast("`${1} ${2}`", vec![int!(1), str!(" "), int!(2)]);
     }
 
@@ -202,30 +203,26 @@ mod tests {
 
     #[test]
     fn test_template_with_complex_expression() {
-        // Original expectation, which might be misaligned with parser output if 'if' and '>' become NativeObjects
-        // If parser produces NativeObject(If) etc., this test will fail, and needs proper test macros (branch!, greater!)
-        // or public constructors for Call to assert the correct NativeObject structure.
         assert_template_ast(
             "`value: ${if x > 0 then y else z}`",
             vec![
                 str!("value: "),
-                // Expecting NativeObject for 'if' and '>'
-                Value::OpCall(Arc::new(Call::new(vec![
-                    Value::NativeObject(Arc::new(Box::new(If::stub()))), // if
-                    Value::OpCall(Arc::new(Call::new(vec![
-                        // condition: x > 0
-                        Value::NativeObject(Arc::new(Box::new(Greater::stub()))), // >
+                Value::OpCall(Arc::new(crate::script::Call::new(vec![
+                    // Path to Call
+                    Value::NativeObject(Arc::new(Box::new(If::stub()))),
+                    Value::OpCall(Arc::new(crate::script::Call::new(vec![
+                        // Path to Call
+                        Value::NativeObject(Arc::new(Box::new(Greater::stub()))),
                         id!("x"),
                         int!(0),
                     ]))),
-                    id!("y"), // then branch
-                    id!("z"), // else branch
+                    id!("y"),
+                    id!("z"),
                 ]))),
             ],
         );
     }
 
-    // Error Conditions
     #[test]
     fn test_unterminated_template_literal() {
         assert_parse_template_error("`abc");
@@ -233,27 +230,22 @@ mod tests {
 
     #[test]
     fn test_unterminated_expression_in_template() {
-        assert_parse_template_error("`${1+2`"); // Missing closing } for expression
-        assert_parse_template_error("`${1+`"); // Missing operand and closing }
+        assert_parse_template_error("`${1+2`");
+        assert_parse_template_error("`${1+`");
     }
 
     #[test]
     fn test_invalid_escape_sequence_in_template() {
-        // Note: \x is not a supported escape in the template string parser
         assert_parse_template_error("`\\x`");
     }
 
     #[test]
     fn test_syntax_error_in_embedded_expression() {
-        assert_parse_template_error("`${1 + * 2}`"); // Invalid operator usage inside ${}
+        assert_parse_template_error("`${1 + * 2}`");
     }
 
     #[test]
     fn test_dangling_dollar_in_template() {
-        // A single '$' not followed by '{' should be treated as a literal '$' or error.
-        // Current parser might treat '$' as start of an escape if not part of ${expr}
-        // The provided `parse_escaped_char` for templates includes `value('$', char('$'))`
-        // which means `\$` is an escaped dollar. A raw `$` not part of `${` is literal.
         assert_template_ast(
             "`hello $world`",
             vec![str!("hello "), str!("$"), str!("world")],
@@ -261,12 +253,8 @@ mod tests {
         assert_template_ast("`$`", vec![str!("$")]);
     }
 
-    // Edge Cases
     #[test]
     fn test_escaped_backticks_and_dollar_signs() {
-        // `\` is escaped as `\\`
-        // `\${value}` should result in literal "${value}" if $ is only special before {
-        // parse_escaped_char in template.rs has: value('$', char('$')), so `\$` -> `$`
         assert_template_ast(
             "`\\`\\${value}`",
             vec![str!("`"), str!("$"), str!("{value}")],
@@ -275,15 +263,11 @@ mod tests {
 
     #[test]
     fn test_template_with_empty_expression_placeholder() {
-        // `${}` is likely a syntax error for the inner expression parser (op_0)
         assert_parse_template_error("`${}`");
     }
 
     #[test]
     fn test_nested_templates_parsed_correctly_by_main_parser() {
-        // This relies on the main parser invoking the template parser recursively.
-        // The `parse_template` function itself doesn't directly handle nesting,
-        // but the `op_0` called inside `${...}` can parse another template.
         assert_template_ast(
             "`outer: ${`inner: ${x}`}`",
             vec![str!("outer: "), strcat!(array!(str!("inner: "), id!("x")))],
