+++ milu/src/script/types.rs
@@ -1,0 +1,61 @@
+use super::traits::NativeObjectRef;
+use std::fmt::Display;
+use std::sync::Arc; // Assuming NativeObjectRef will be in traits.rs
+
+#[derive(Debug, Eq, Clone)]
+pub enum Type {
+    String,
+    Integer,
+    Boolean,
+    Array(Box<Type>),
+    Tuple(Vec<Type>),
+    NativeObject(Arc<NativeObjectRef>),
+    Any, //native only
+}
+
+impl Type {
+    pub fn array_of(t: Self) -> Self {
+        Self::Array(Box::new(t))
+    }
+    pub fn tuple_of(t: Vec<Self>) -> Self {
+        Self::Tuple(t)
+    }
+}
+
+impl PartialEq for Type {
+    fn eq(&self, other: &Type) -> bool {
+        use Type::*;
+        match (self, other) {
+            (Any, _) => true,
+            (_, Any) => true,
+            (String, String) => true,
+            (Integer, Integer) => true,
+            (Boolean, Boolean) => true,
+            (Array(a), Array(b)) => a == b,
+            (Tuple(a), Tuple(b)) => a == b,
+            (NativeObject(a), NativeObject(b)) => a == b,
+            _ => false,
+        }
+    }
+}
+
+impl Display for Type {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Self::String => write!(f, "string"),
+            Self::Integer => write!(f, "integer"),
+            Self::Boolean => write!(f, "boolean"),
+            Self::Array(a) => write!(f, "[{}]", a),
+            Self::Tuple(t) => write!(
+                f,
+                "({})",
+                t.iter()
+                    .map(ToString::to_string)
+                    .collect::<Vec<_>>()
+                    .join(",")
+            ),
+            Self::NativeObject(x) => write!(f, "native@{:x}", x.gen_hash()),
+            Self::Any => write!(f, "any"),
+        }
+    }
+}
